Subject: initialization of std::string_view with null pointer
When switching to C++17, absl::string_view implicitly uses
std::string_view. Initialization of string_view with nullptr was
allowed in abseil implementation of string_view but the same is not
true in std. https://github.com/google/cel-cpp/pull/45 fixes majority
of these instances. But, in this instance, it has been left out.

Signed-off-by: Yifan Yang <needyyang@google.com>