Introduce CelTypeRegistry for tracking type identifiers with updates in type resolution.
Type and enum constant values may be shadowed by variables with the same
name provided within an Activation in order to preserve backward compatibility
with potential existing usages of the library.

This change makes it possible to find core CEL type names as identifiers, and
if the `enable_qualified_type_identifiers` option is enabled, then qualified
names which appear within Select expressions can be resolved to types which
have either been registered with the CelTypeRegistry or to protobuf type names
which have been linked into the binary.

PiperOrigin-RevId: 353680095