Handle the case where a macro call may have macros nested in its arguments, such as
x.filter(y, has(y.a) && has(y.b))

Because the first arg to filter is the call expr &&, this arg will not be found in macro_calls and the entire expression will be stored in the map, including the macro expansion for has.

These macros are already present in macro calls, so adding a macro call can recursively look through its arguments for macros and replace them with the proper ID from macro calls.

PiperOrigin-RevId: 399285474