Ensure comprehension accumulation cannot allocate memory exponentially for hand-crafted ASTs.
It is possible for an accumulation variable in a comprehension to be
referenced more than once in the `loop_step` of a comprehension. Sometimes
this is fine, as is the case within the ternary operation `?:` which either
appends to the accumulation variable or returns the existing value. In other
cases, such as `+`, it is possible for the value to grow and allocate
exponentially, e.g.

```
    iter_range: [0, 1, 2, 3, 4, 5, 6, 7, 8, 9]
    accu_init: ["hello"],
    accu_var: "accu",
    loop_step: accu + accu
    loop_condition: accu == accu
```

In the example, the size of the `accu` grows to 2^iter_range elements.

Within parser generated comprehensions generated by the core CEL libraries,
such expressions are impossible. But in hand-rolled ASTs, it is very easy to
induce memory explosion.

PiperOrigin-RevId: 404640801