Add CEL/C++ support for hermetic descriptor pools
So far, CEL C++ always used only the generated descriptor pool (and message
factory). This means only protos that are compiled into the binary using CEL
can be used.

In our scenario, we have a database providing certain entities that can be
added in our system at run-time and which use proto to describe output data.
We want to use CEL to formulate expressions over this output. Since we know
only at run-time what protos these are, we need to create a hermtic descriptor
pool and dynamic message factory for that transitively closed file descriptor
set.

This CL adds the capability to CEL/C++ to use a custom descriptor poold and
message factory. From the users' perspective this is entirely optional and
the generated descriptor pool and message factory will be used if not
explicitly overridden.

PiperOrigin-RevId: 427701926