Fix nullability annotation of `ReferenceCountingMemoryManager::Allocate()`.
The declaration didn't have an annotation; the definition is annotated as
returning nonnull, but it contains a `return nullptr`.

For the time being, I'm fixing this by removing the annotation on the definition
(thereby making it consistent with the declaration). Code owners may want to
follow up by marking both the definition and declaration as returning nullable
if desired.

PiperOrigin-RevId: 639685981