Breaking change: Use dev.cel (canonical) protos instead of google.api.expr
Historically C++ was implemented using the unversioned protobuf descriptors. It eventually included a mix of the versioned variant. When moving to OSS, unversioned protobuf descriptors were rewritten to the versioned variant unconditionally. This has created a bit of a conundrum, now that their is an unversioned canonical protobuf descriptor in cel-spec. We have decided to bite the bullet and break the world. The versioned and unversioned protobuf descriptors are wire compatible, so if you run into issues you can serialize and parse to convert between the two.

PiperOrigin-RevId: 689865345