Silence gcc overloaded-virtual warning in FieldBackedMapImpl
gcc complains about FieldBackedMapImpl::ListKeys when overloaded-virtual
check is enabled. When using -Werror, warnings like this are converted
to errors breaking builds.

That's what happened to gcc build of Envoy a while back. The issue in
Envoy was fixed by introducing a patch for cel-cpp that was applied to
the source during Envoy build (see
https://github.com/envoyproxy/envoy/pull/31814).

Unfortunately during cel-cpp version update in Envoy instead of
updating the patch, dropping the parts that are not needed anymore, we
dropped the whole cep-cpp patch all together in
https://github.com/envoyproxy/envoy/pull/36661. And now gcc can't build
Envoy again.

This change makes the change to silence the gcc warning in the cel-cpp
codebase, which hopefully will be a bit more durable than maintaing a
patch on the Envoy side.

Signed-off-by: Mikhail Krinkin <krinkin.m.u@gmail.com>