Add specialized impls for !_ and @not_strictly_false
Adds custom evaluator steps for !_ and @not_strictly_false as an alternative
for the extension function based implementation. These tend to be called at high
frequency (e.g. .exists and .all, predicates), so they benefit from the reduced overhead.

This is optional since it prevents disabling or extending the operators.

PiperOrigin-RevId: 716319261