Default to using specialized implementation for some builtins.
Applies to !/@not_strictly_false/==/!=/@in.

This avoids some of the overhead for function dispatch at the cost of duplicating a subset of that logic and restricting the ability to customize them.

Note: the fast builtins assume heterogeneous equality behavior. If not enabled, the runtime will fall back to the provided equals functions, instead of the new builtins step.
PiperOrigin-RevId: 717979767