Add AddUnknownPattern() and AddMissingPattern().
The other activation functions can be used incrementally, but SetUnknownPatterns() and SetMissingPatterns() cannot. This makes it a bit annoying to use, since you can't have a wrapper function that either binds an attribute or add an unknown pattern, that you could for example call in a loop.

Since attribute patterns are just a vector anyway, it is easy to make it support incremental additions.

PiperOrigin-RevId: 748766532