Avoid nesting status macros
In MSVC, when there are nested macro invocations, the value of __LINE__ will be the same across the nested invocations. This causes temporaries from multiple nested invocations of CEL_ASSIGN_OR_RETURN to conflict with each-other.

Thankfully, this issue seems to only occur in one place in the entire codebase, and it probably won't recur too often.